Перенес логику в App.vue
