+++ src/frontend/src/layouts/AppLayoutMain.vue
@@ -1,6 +1,6 @@
 ﻿<template>
-  <div class="">
-    <AppLayoutHeader :total-price="totalPrice" />
+  <div>
+    <AppLayoutHeader :total-price="totalPrice" :user="user" />
     <slot />
   </div>
 </template>
@@ -9,13 +9,17 @@
 import AppLayoutHeader from "@/layouts/AppLayoutHeader";
 
 export default {
-  name: "AppLayoutDefault",
+  name: "AppLayoutMain",
   components: {
     AppLayoutHeader,
   },
   props: {
     totalPrice: {
       type: Number,
+      required: true,
+    },
+    user: {
+      type: Object,
       required: true,
     },
   },
