+++ src/frontend/src/views/Index.vue
@@ -20,11 +20,11 @@
         />
 
         <BuilderPizzaContent
-          :price="totalPizzaPrice"
+          :price="totalPrice"
           :selected-items="selectedItems"
           :fillings="selectedFillings"
           :name="pizza.name"
-          @changeNamePizza="changeNamePizza"
+          @changeNamePizza="$emit('inputName', $event)"
           @dropFilling="dropFilling"
         />
       </div>
@@ -33,9 +33,6 @@
 </template>
 
 <script>
-import pizzaData from "@/static/pizza.json";
-import { getPizzaValues } from "@/common/helpers";
-import { MAX_COUNT_TYPE_INGREDIENT } from "@/common/constants";
 import BuilderDoughSelector from "@/modules/builder/components/BuilderDoughSelector";
 import BuilderSizeSelector from "@/modules/builder/components/BuilderSizeSelector";
 import BuilderIngredientsSelector from "@/modules/builder/components/BuilderIngredientsSelector";
@@ -49,89 +46,33 @@ export default {
     BuilderSizeSelector,
     BuilderDoughSelector,
   },
-  data() {
-    return {
-      pizza: getPizzaValues(pizzaData),
-    };
+  props: {
+    pizza: {
+      type: Object,
+      required: true,
+    },
+    selectedItems: {
+      type: Object,
+      required: true,
+    },
+    selectedFillings: {
+      type: Array,
+      required: true,
+    },
+    totalPrice: {
+      type: Number,
+      default: 0,
+    },
   },
   methods: {
     clickSelectorItem: function (selector) {
-      this.pizza[selector.type] = this.pizza[selector.type].map((elem) => ({
-        ...elem,
-        checked: selector.id === elem.id,
-      }));
+      this.$emit("clickSelectorItem", selector);
     },
     clickButtonItemCounter: function (filling) {
-      this.pizza.fillings = this.pizza.fillings.map((elem) => {
-        if (elem.id !== filling.id) return elem;
-
-        const count =
-          filling.typeClick === "increase" ? ++filling.count : --filling.count;
-
-        return {
-          ...elem,
-          count,
-          permissions: {
-            decrease: count > 0,
-            increase: count < MAX_COUNT_TYPE_INGREDIENT,
-          },
-        };
-      });
-    },
-    changeNamePizza(event) {
-      this.pizza.name = event.target.value;
+      this.$emit("clickButtonItemCounter", filling);
     },
     dropFilling(filling) {
-      this.clickButtonItemCounter({ ...filling, typeClick: "increase" });
-    },
-  },
-  computed: {
-    selectedItems() {
-      return {
-        dough: this.selectedDough,
-        size: this.selectedSize,
-        sauce: this.selectedSauce,
-        fillings: this.selectedFillings,
-      };
-    },
-    selectedDough() {
-      return this.pizza.doughs.find((elem) => {
-        return elem.checked === true;
-      });
-    },
-    selectedSize() {
-      return this.pizza.sizes.find((elem) => {
-        return elem.checked === true;
-      });
-    },
-    selectedSauce() {
-      return this.pizza.sauces.find((elem) => {
-        return elem.checked === true;
-      });
-    },
-    selectedFillings() {
-      return this.pizza.fillings.filter((item) => item.count);
-    },
-    totalPizzaPrice() {
-      const fillingsPrice = this.selectedItems.fillings.reduce((acc, elem) => {
-        const { count, price } = elem;
-        return acc + count * price;
-      }, 0);
-
-      return (
-        (this.selectedItems.dough.price +
-          this.selectedItems.sauce.price +
-          fillingsPrice) *
-        this.selectedItems.size.multiplier
-      );
-    },
-  },
-  watch: {
-    totalPizzaPrice: {
-      immediate: true,
-      handler(totalPrice) {
-        this.$emit("price", totalPrice);
-      },
+      this.$emit("dropFilling", filling);
     },
   },
 };
